Minor update to allow registration of system Classes
In this version a class registered on system class list shall be enabled
to set the isInternal flag of the SvCore via setIsInternal(boolean)
method. 

Minor tweaks to moving utility methods for getting class name to the
SvUtil class. Also modification of the SvConf to handle the new svarog
parameter