Svarog Cluster bugfix and release 3.2.13
Svarog 3.2.13 changes:

In svarog.properties, users can attach startup or shutdown hooks.
These are called when the node joins a cluster or when the node
parts a cluster or shuts down.

"svarog.shutdown.hook";
"svarog.startup.hook";

Also, all svarog methods can now raise exception with code:
"system.error.cluster_inactive" which signifies that the node is
in limbo and is not part of any cluster